sanphelp_microservice/comment_service/requirements.txt
amqp==1.4.8
anyjson==0.3.3
Django==1.8.4
django-braces==1.8.1
django-cors-headers==1.1.0
django-extensions==1.5.7
django-filter==0.11.0
django-push-notifications==1.4.1
django-rest-swagger==0.3.5
django-storages==1.1.8
djangorestframework==3.2.4
drfdocs==0.0.6
enum34==1.0.4
html5lib==0.9999999
Markdown==2.6.2
meld3==1.0.2
MySQL-python==1.2.5
pdfkit==0.5.0
pyparsing==1.5.7
PyPDF2==1.26.0
pytz==2015.7
requests==2.7.0
six==1.9.0
xhtml2pdf==0.0.6
xmltodict==0.10.1

